integração get trip details
